src/services/multimodal/docling/requirements.txt
# Docling Multimodal Service Dependencies
# Python 3.11 required

# Core - Docling (all document processing)
docling

# OCR Support - RapidOCR (based on PaddleOCR, better Chinese recognition)
rapidocr-onnxruntime
onnxruntime

# Audio Transcription
openai-whisper

# Web Framework
fastapi
uvicorn
